pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["Glenn Elliron Clark <Elliron.Clark@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework-simplejwt = "^4.8.0"
Django = "^3.2.7"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
